+++ clang-tools-extra/clang-tidy/llvm/LLVMTidyModule.cpp
@@ -18,7 +18,7 @@
 #include "PreferRegisterOverUnsignedCheck.h"
 #include "PreferStaticOverAnonymousNamespaceCheck.h"
 #include "RedundantCastingCheck.h"
-#include "RegexCheck.h"
+#include "InvalidRegexPatternCheck.h"
 #include "TwineLocalCheck.h"
 #include "TypeSwitchCaseTypesCheck.h"
 #include "UseNewMLIROpBuilderCheck.h"
@@ -49,7 +49,7 @@ public:
         "llvm-qualified-auto");
     CheckFactories.registerCheck<RedundantCastingCheck>(
         "llvm-redundant-casting");
-    CheckFactories.registerCheck<RegexCheck>("llvm-regex");
+    CheckFactories.registerCheck<InvalidRegexPatternCheck>("llvm-invalid-regex-pattern");
     CheckFactories.registerCheck<TwineLocalCheck>("llvm-twine-local");
     CheckFactories.registerCheck<TypeSwitchCaseTypesCheck>(
         "llvm-type-switch-case-types");